+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming
 
 - feat(rpc): add a proto generator which parses actions and events files and generates proto files for them
+- feat(rpc): add `rpc` service with `dispatch` method to let external services dispatch actions and events to the redux bus
 
 ## Version 0.16.2